feat: add a seperate component to get back old v2.0.0 and prior data
